MXHXMacroResolver: fix recursive calls to resolveQname() and resolveTag()
Certain things that the resolver does can trigger other build macros to run,
which can cause the resolver to call itself recursively. when that happens,
certain symbol fields may not be fully populated yet. to fix this, check if
anything is pending before returning from calls to the public API and finish
resolving if there is.